Add .eslint/.prettier; Reformat *.js files; Remove unused files

+++ contracts/Migrations.sol
@@ -9,10 +9,9 @@
 
 pragma solidity ^0.5.17;
 
-
 contract Migrations {
   address public owner;
-  uint public last_completed_migration;
+  uint256 public last_completed_migration;
 
   constructor() public {
     owner = msg.sender;
@@ -22,7 +21,7 @@ contract Migrations {
     if (msg.sender == owner) _;
   }
 
-  function setCompleted(uint completed) public restricted {
+  function setCompleted(uint256 completed) public restricted {
     last_completed_migration = completed;
   }
